% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave_base64}
\alias{ggsave_base64}
\title{Save a ggplot2 plot as a base64 encoded string}
\usage{
ggsave_base64(
  plot = ggplot2::last_plot,
  format = c("png", "jpeg", "svg"),
  width = 500,
  height = 500,
  ...
)
}
\arguments{
\item{plot}{A ggplot2 plot object.}

\item{format}{The format to save the plot in. One of \code{"png"},
\code{"jpeg"}, or \code{"svg"}.}

\item{width}{The width of the plot in pixels.}

\item{height}{The height of the plot in pixels.}

\item{...}{Additional arguments to pass to \code{\link[ggplot2]{ggsave}}.}
}
\description{
Save a ggplot2 plot as a base64 encoded string
}
